fix: share oobi instead of name and identifier
